requirements.txt
flask
Flask-Cors 
waitress
regex
Flask-RESTful
mysql-connector
uvicorn
fastapi